miniscript: add CONSENSUS and SANE consts to all contexts
We are going to move all of the logic currently ad-hoc spread across the
various "global"/"local" "consensus"/"policy" checks and other related
checks into a single validate() method implemented on Miniscript, Policy,
Semantic and possibly other types.

This validate() method will have an understanding of which checks are
"local" (i.e. computable from a single node) and which are "global" (i.e.
require looking at the whole script). This distinction is not meaningful
to an end user and it's hard for even developers to keep track of. For
a user, all the checks are just checks, and their knob belongs in the
big ValidationParams structure.