miniscript: rename parse_* functions to decode_*
We should backport this one to 12.x (with a deprecation). Would also be
open to doing a deprecation here in master. But this has been bugging me
for years, and also it's much less common to decode miniscript from
script than we imagined back in 2018.

Provides symmetry with encode() and consistency with the docs (though
this PR does not fix up the docs).