miniscript: replace decode_with_ext(., Ext::allow_all) with decode_consensus
Currently we have decode() and decode_insane() (which were both called
parse_* before PR #845). In practice though, throughout the codebase we
see that when decoding from script we want to relax more than sanity.

In particular, we frequently with ExtParams::allow_all, which beyond
relaxing the sanity rules, also allows raw pubkeyhashes. We do this in
the script interpreter and in PSBT, on the basis that once we're working
with a script, we should just deal with whatever the library is capable
of dealing with. (Is this the right decision? IMO yes. It could be
argued. But regardless, it's the decision we've made for the past many
versions of rust-miniscript.)

I also propose to backport this with deprecation, telling users that if
they really want parse_insane, they now need to call decode_with_ext and
specify ExtParams::insane. But if their goal was "parse anything that
might be allowed", they actually want decode_consensus.

There is now an asymmetry between parsing from string and parsing from
script: with strings we have parse() and parse_insane(). With script we
have decode() and decode_consensus(). I think this is correct, and
reflects the fact that somebody "breaking the rules" with a string is
likely trying to use syntactically valid Miniscripts without the library
whining at him, while somebody "breaking the rules" with a Script is
probably trying to deal with some immutable on-chain thing and wants the
library to work if at-all possible.

Right now the distinction is simply "do we support raw pkh or not" but I
think we could accept more-or-less arbitrary extensions to Miniscript in
this library that worked the same way (allowed when decoded from script
but disallowed from strings since there's no serialization).