miniscript: replace from_str_ext with from_str_with_validation_params
This moves us toward being able to drop ExtParams. It also eliminates
the NonTopLevel error variant, which is now ValidationError::NonBase
and controllable via ValidationParams::allow_non_b.

Also tweaks the "hack" in Tr::from_tree to be more sensible, though it's
still a hack. A later commit will change the FromTree trait to incorporate
validation and then the inconsistency this hack reflects will go away.

Leaves the from_str_insane method, with cleaned up doccomment. Probably
we should deprecate this since I don't like the term "insane", but we
can do that in a followup if that's what we want to do.