remove all sanity_check and ext_check methods from the library
Well, there are sanity checks in the type system (which are private and
make sense) and a sanity_check method in the PSBT module that I didn't
look at. But the rest are gone now.

These should really never have existed. At this point in this commit
series we are pretty close to a point where they serve no purpose at
all; miniscripts constructed with from_ast, from_str and decode are
all sanity-checked at construction with the same set of checks.

We haven't yet attempted to apply these validation parameters to either
of the policy types, or gone through all the non-miniscript descriptor
types, but we'll get there. The next series of PRs are going to overhaul
all the error types in the library based on this validation framework.

I considered deprecating these but we are also deleting the AnalysisError
error that it returns. We could reintroduce them with a return type like
Result<(), Infallible>, which in a decent language would be non-breaking,
but Rust is indecent and this wouldn't really help callers.

Also deletes AnalysisError, ExtParams, and the Analysis variant of the
top-level Error enum!