descriptor: use new validation in the other constructors
This is another big commit that has a bunch of behavior changes of the
form "now we enforce sanity rules where we didn't use to". This one lets
us delete a huge chunk of context.rs which is nice.

This eliminates the NonStandardBareScript error variant, but no tests
fail and I didn't replace the functionality. If I open a PR with this
commit in this state, we need to file an issue to restore this and to
add tests for it.