sortedmulti: take Thresh as input to constructors
We currently take a k value and a vector as input to the constructor to
the SortedMultiVec type, and propagate this out to all the sortedmulti
constructors. This means that we have to construct the threshold inside
the method and then return errors outward.

It's more flexible, though potentially a bit more annoying, to make the
user construct the Threshold. Then they can deal with the error. This
means that the descriptor constructors *only* return a validation error,
which will let us tighten the Result type in a later commit.