descriptor: use new validation in constructors for pkh descriptors
Update sortedmult, pkh, wpkh to the new validation framework. These
constructors now return a ValidationError rather than an Error, which
is a more precise type. The constructors also actually do validation;
in particular sortedmulti now does a duplicate pubkey check where it
didn't use to.

The theme over these next several commits is going to be "we now do
checks where we didn't use to". It may be that we get pushback from
users over this. In that case we need to adapt somehow but at least
we can do it in a principled way, rather than the current "there are
a ton of different ways to check and they're all randomly half
implemented".