plan: add some clarifying docs
"The size in bytes of the script sig that" is ambiguous and could
refer to the actual scriptSig, exlcuding the varint prefix that
indicates its size.

The code comment in witness_size is promoted to the docstring, as its
helpful to anyone trying to compute the right size of a transaction.