psbt: make PsbtInputSatisfier internals private
We don't need access to the internals anywhere, and this is just a POD
type with two Copy objects inside so the user can just keep them around
after construction if they need access.

Add an accessor for the PSBT anyway since it might be convenient.

Making these fields private does not break compilation at all.