fuzz/generate-files: support jujutsu
Generally we try not to add special-case logic for weird tools, but this
is a a really simple change and I think that jujutsu is going to become
pretty popular among rust-bitcoin developers.

This is some repeated code which looks like it should be factored out
into a utility function, but we can't really do that because we need to
figure out the workspace root before we can even locate utility functions.