descriptor/key.rs: clean up error matching
This uses `Self::XYZ` when matching on errors, rather than writing
out the full name or using wildcard imports (which are dangerous
because the compiler may misinterpret typos as being variable
names) (although there will be a warning and CI should fail).

This also collapses a bunch of fmt impls to only write the top-level
error, in cases where we weren't adding any useful context. (We
should revisit these and produce rich errors which have line number
information and so forth; and I believe the correct way to describe
the errors is the opposite of what I've done, where we are supposed
to provide context but NOT the underlying error which can be accessed
by the .source() method. But this simplifies the code without making
things worse and we can do a proper fix later.)