clippy: fix most lints
This, and the next commit, are not organized in any way except that I
did a new commit partway to try to make review a bit easier. There
were too many lints for me to split them all up and do them in a nice
PR.

This changes `OrdF64::partial_cmp` to be consistent with `OrdF64::cmp`
which is a bugfix and an observable behavior change (but not an easy
one to observe). All other changes are just syntax stuff.