ci: set all runners to ubuntu-latest
It may be that ubuntu-latest breaks us, but we're not doing anything
particularly exotic so it seems unlikely.

Meanwhile, our use of ubuntu-20.04 HAS broken us, because that runner
is no longer supported.

Meanwhile also update a bunch of our action versions to v4, since some
of the v2s have become deprecated and removed. (Hopefully this won't
happen very often.)