clippy: non-reference stuff
As in #835, I changed a test function that used 'A'..'Z' to use an
inclusive range, which is behavior-changing, and changed the impl
of OrdF64::partial_cmp to be correct.

Everything else is just syntactic.

I apologize for structuring this PR very differently from #835 so you
can't really range-diff. I tried rebasing and there were so many
conflicts that I simply redid the whole thing, and since I had some
experience I made some different decisions about what order to take
things in.