test: delete bitcoind tests
Sigh, after all the work I did to get a lockfile for the previous commit,
it turns out that all the bitcoind tests were disabled anyway, with the
explanation "breaking change in bitcoind".

This commit deletes them entirely. We have no plans to fix them. If we
ever want to regression test miniscript 9.x, we should do it externally
by adding a fuzztest in master which pulls it in and tries to parse
stuff. Then it will be actively exercised and maintained and not require
ancient broken dependencies.

This commit is all red EXCEPT a change to a test in src/miniscript/mod.rs.
What's happening here is that we had one test (asserting on the string
representation of errors from rust-bitcoin) which fails when the std
feature is off. As it turns out, `bitcoind` unconditionally turns on that
feature. So by having `bitcoind` as a dev-dependency it was impossible
for us to ever test the nostd feature properly. Sigh.