miniscript/iter: make test module non-pub
We have a `pub mod tests` module which is commented as being public
"in case dependent libraries want to use these functions". However,
it's also gated by #[cfg(test)]. I am not aware of any way that
dependent libraries could compile *this* one with cfg(test) on. So
this module was, in effect, never public.

Make it non-pub to make things clearer, and also to silence all the
"missing docs" warnings that appear when running clippy with --all-targets.

This is NOT an API-breaking change even though it removes a public module,
because the module was never actually usable.