docs: fix broken links etc
Remove the deprecated (well, renamed) "allow broken intradoc links" cfg
flag, which exposes a ton of broken doc links. Mostly just uses of []
or <> that aren't in backticks. Fix them all.

The one morally questionable change is to change two doclink references
to Policy::enumerate_policy_tree to no longer be doclinks. The problem
is that this function only exists when the `compiler` flag is on. In
master in ebaa31bca7e80f02751e3435907b3aad99d74168 we just removed
these references as well, but arguably we should have fixed them.

Anyway in 9.x we're certainly not going to bend over backward to fix
stuff like this.