key: fix DefiniteDescriptorKey construction from key with hardened step
Our current API allows constructing a DefiniteDescriptorKey with hardened steps,
which means that you can't actually derive a public key. However, the point of
this type is to provide a DescriptorPublicKey which implements the ToPublicKey
trait.

This is a backport, so it introduces the `has_hardened_steps` helper function
but does *not* make it pub, in the interest of avoiding gratuitious API changes.