psbt: remove derives from UtxoUpdateError and OutputUpdateError
Going forward, we should only derive Debug and whichever of Clone,
PartialEq and Eq are possible.

PartialEq is very useful for doing assertions, and Clone is often
useful because it allows Results to be cloned. But the other traits
are uncommon to implement on error types, have an unclear meaning,
and limit our ability to compose types without breaking changes.

Also adds a bunch of Eq bounds to error types that implement PartialEq.
There is almost never a reason to implement PartialEq but not Eq.

In the next commit I'm gonna replace ConversionError, an old error
which derives all these things, with the new NonDefiniteKeyError,
which will prevent these derives from working on UtxoUpdateError.