miniscript: use downcasting logic for computing a taproot leaf hash
Right now we have a method `tap_leaf_internal` which unconditionally
computes the "tapleaf" of a Miniscript. If the Miniscript is a Taproot
script, this method is exposed by the public function `tap_leaf`.

This diff inverts `tap_leaf` and `tap_leaf_internal` -- now `tap_leaf`
does the actual calculation (and is implemented on a wider set of
Miniscripts, no longer restricting the key type unnecessarily) and
`tap_leaf_internal` calls `tap_leaf` after downcasting.

The internal method is called by Miniscript::satisfy, which passes it
into the recursive functions of src/miniscript/satisfy.rs, which carefully
avoid using it except when Ctx is equal to Tap. Aside from being wasteful,
this logic is hard to verify the correctness of, and won't survive the
removal of the Ctx parameter.

This commit just sets the TapLeafHash to all zeroes in the case that it's
unused. The next commit will replace it with an Option<TapLeafHash>, but
I didn't want to do that in this commit because it increased the diff a
lot by forcing a bunch of API changes.

The goal of this commit is just to introduce the downcasting logic, which
uses unsafe code and ought to be reviewed carefully.