miniscript/satisfy: move pk_k logic into helper function
For every fragment, we are going to create a short helper function which
computes the satisfaction and dissatisfaction for that fragment. We create
these together because in general satisfactions and dissatisfactions are
defined in terms of each other:

* Several satisfactions (e.g. `or_b`) require dissatisfactions to create
* Satisfaction of `and_v` requires the satisfaction of its V child to create

When we un-recursify these functions we will therefore need to compute both
satisfactions and dissatisfactions together. (The result will be less
efficient than the recursive version, since it will compute both sats and
dissats for every fragment rather than computing just the data that's
needed. But it should be easier to follow.)

The next commit will do this for several more fragments. But I am doing it
for pk_k in a commit by itself so you can see the process.