miniscript/satisfy: rewrite satifier to be non-recursive
This rewrites the main satisfy function, which I apologize for, and it's not
even a simple code move because I combined satisfy_ and dissatisfy_helper,
and had to change the code to pop from the stack rather than doing recursive
calls.

Furthermore, I suspect that this code will be slower in some cases, though
unlikelily in a way that anybody would ever notice, because it computes
both the satisfaction and dissatisfaction for every single node, not just
the ones that are needed. (Note: it was always the case that, when
satisfying, that a satisfaction was attempted for every single node, and
that a dissatisfaction would be attempted for most nodes contained in any
disjunction. So probably satisfaction is not meaningfully slowed, and
importantly, we aren't making any more calls to the satisfier object,
which might be an API to some slow USB device.)

HOWEVER, the resulting code is much smaller and cleaner. I found a bug, I
think (noted with a FIXME to add a test). I deleted over twice as much
code as I added, and in particular eliminated tons of extremely noisy
generics related to mall/non-mall function pointers. (The next commits
will remove even more generics.)