fuzz: switch from honggfuzz to cargo-fuzz
honggfuzz is increasingly hard for me to maintain in my local CI. It
requires an exact match between the version of honggfuzz in the lockfile
and the binary installed in the nix derivation, and these exact versions
come with specific requirements about which rustc versions are allowable
(and in particular, honggfuzz 0.5.56 does not work with rustc 1.74.0,
which is the MSRV for rust-bitcoin). I also have to bend over backward
to set compiler flags and make source directories writable.

cargo-fuzz, meanwhile, requires nightly to *run* the fuzz tests, but does
not require it for just running unit tests or building. (The latest
version of libfuzzer-sys uses once_cell and requires a higher MSRV
than we have, but we don't need to use the very latest version.)